Now attempting to do message formatting for MS Teams - refactored to reuse much of the slack work
